# backend/src/swagger.yaml
openapi: "3.0.3"
info:
  title: Plotra API
  description: 소설 창작 AI 보조 플랫폼 - REST API
  version: "1.0.0"
servers:
  - url: http://localhost:4000
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, nickname]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 1 }
                nickname: { type: string }
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/UserPublic" }
                  token: { type: string, description: "Encrypted JWT token" }
        "400":
          description: Missing fields
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string } } }
        "409":
          description: Email already registered
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string } } }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/UserPublic" }
                  token: { type: string, description: "Encrypted JWT token" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string } } }
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      operationId: getMe
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPublic" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string } } }

  # ─── Health ───
...
  /api/projects/{projectId}/chapters:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Chapters]
      summary: List chapters
      operationId: listChapters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Chapter" }
    post:
      tags: [Chapters]
      summary: Create chapter
      operationId: createChapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                content: { type: string }
                summary: { type: string }
                sortOrder: { type: integer }
                metadata: { type: object }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chapter" }
  /api/projects/{projectId}/chapters/{id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Chapters]
      summary: Get chapter
      operationId: getChapter
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chapter" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Chapters]
      summary: Update chapter
      operationId: updateChapter
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                content: { type: string }
                summary: { type: string }
                sortOrder: { type: integer }
                metadata: { type: object }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chapter" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Chapters]
      summary: Delete chapter (soft)
      operationId: deleteChapter
      responses:
        "200": { $ref: "#/components/responses/Deleted" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Characters ───
...
  /api/projects/{projectId}/plots/{id}/chapters:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Plot Chapters]
      summary: List linked chapters
      operationId: listPlotItemChapters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PlotItemChapter" }
    post:
      tags: [Plot Chapters]
      summary: Link chapter to plot item
      operationId: createPlotItemChapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chapterId]
              properties:
                chapterId: { type: string, format: uuid }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlotItemChapter" }
  /api/projects/{projectId}/plots/{id}/chapters/{linkId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
      - $ref: "#/components/parameters/LinkId"
    delete:
      tags: [Plot Chapters]
      summary: Unlink chapter from plot item
      operationId: deletePlotItemChapter
      responses:
        "200": { $ref: "#/components/responses/Deleted" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Plot ↔ Characters ───
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "Encrypted JWT token"
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    LinkId:
      name: linkId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    Deleted:
      description: Deleted
      content:
        application/json:
          schema:
            type: object
            properties:
              deleted: { type: boolean }
  schemas:
    UserPublic:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        nickname: { type: string }
        createdAt: { type: string, format: date-time }
...
    Chapter:
      type: object
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        title: { type: string }
        content: { type: string }
        summary: { type: string, nullable: true }
        sortOrder: { type: integer }
        metadata: { type: object }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
...
    PlotItemChapter:
      type: object
      properties:
        id: { type: string, format: uuid }
        plotItemId: { type: string, format: uuid }
        chapterId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
...